Fix lint, move package to a better name
